pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xapi-client"
version = "0.1.0"
description = "Typed, async-first Python client for xAPI Learning Record Stores"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Pranav Shastri" }]
keywords = ["xapi", "lrs", "learning-record-store", "tin-can", "e-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.5,<3.0",
    "python-dateutil>=2.8,<3.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.urls]
Homepage = "https://github.com/shastripranav/xapi-client"
Issues = "https://github.com/shastripranav/xapi-client/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/xapi_client"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
